data_api_service/requirements.txt
fastapi
uvicorn
pymongo
bcrypt
pydantic[email]
pyjwt